Change syntax to simplify parser
Every source file is now a single expression. The expression
is bound to an identifier of the same name with the source file.
This is pretty big limitation but will hopefully allow me to make
some progress in more interesting parts of the compiler rather than
getting stuck with the intricacies of the parser.